Complete howe task. - MusicPlayer contains list of mixed music genres. - Add list-bean to SpringConfig. - MusicPlayer play random music. - No @Component/Autowired/ComponentScan p.s. Computer not taken into account.
