+++ src/main/java/ru/ostanin/springdemo/Music.java
@@ -2,8 +2,6 @@ package ru.ostanin.springdemo;
 
 import org.springframework.stereotype.Component;
 
-import java.util.List;
-
 public interface Music {
     String getSong();
 }
