+++ src/main/java/ru/ostanin/springdemo/MusicGenre.java
@@ -1,5 +1,5 @@
 package ru.ostanin.springdemo;
 
 public enum MusicGenre {
-    CLASSICAL, ROCK, EDM
+    CLASSICAL, ROCK, EDM, JAZZ
 }
