--- src/main/java/ru/ostanin/springdemo/SpringConfig.java
@@ -1,36 +1,0 @@
-package ru.ostanin.springdemo;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-
-@Configuration
-//@ComponentScan("ru.ostanin.springdemo")
-@PropertySource("classpath:musicPlayer.properties")
-public class SpringConfig {
-    @Bean
-    public ClassicMusic classicMusic() {
-        return new ClassicMusic();
-    }
-
-    @Bean
-    public RockMusic rockMusic() {
-        return new RockMusic();
-    }
-
-    @Bean
-    public EDMMusic edmMusic() {
-        return new EDMMusic();
-    }
-
-    @Bean
-    public MusicPlayer musicPlayer() {
-        return new MusicPlayer(rockMusic(), edmMusic());
-    }
-
-    @Bean
-    public Computer computer() {
-        return new Computer(musicPlayer());
-    }
-}
