+++ src/main/java/ru/ostanin/springdemo/TestSpring.java
@@ -1,7 +1,8 @@
 package ru.ostanin.springdemo;
 
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+import ru.ostanin.springdemo.config.SpringConfig;
+import ru.ostanin.springdemo.genres.ClassicMusic;
 
 public class TestSpring {
     public static void main(String[] args) {
@@ -9,35 +10,7 @@ public class TestSpring {
         AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(SpringConfig.class);
 
         MusicPlayer musicPlayer = context.getBean("musicPlayer", MusicPlayer.class);
-
-        System.out.println(musicPlayer.getName());
-        System.out.println(musicPlayer.getVolume());
-
-        ClassicMusic classicMusic = context.getBean("classicMusic", ClassicMusic.class);
-        ClassicMusic classicMusic1 = context.getBean("classicMusic", ClassicMusic.class);
-
-        System.out.println(classicMusic == classicMusic1);
-        /*Computer computer = context.getBean("computer", Computer.class);
-        System.out.println(computer);*/
-
-        //Music myMusic = context.getBean("electronic", Music.class);
-        //System.out.println(myMusic.getSong());
-
-        //MusicPlayer myPlayer = context.getBean("musicPlayer", MusicPlayer.class);
-        //myPlayer.playFavoriteMusic();
-
-        /*MusicPlayer firstPlayer = context.getBean("musicPlayer", MusicPlayer.class);
-        MusicPlayer secondPlayer = context.getBean("musicPlayer", MusicPlayer.class);
-        firstPlayer.playMusic();
-
-        firstPlayer.setVolume(8);
-        secondPlayer.setName("Temp player");
-
-        System.out.println(firstPlayer.getName());
-        System.out.println(firstPlayer.getVolume());
-
-        System.out.println(secondPlayer.getName());
-        System.out.println(secondPlayer.getVolume());*/
+        System.out.println(musicPlayer.playFavoriteMusic());
 
         context.close();
     }
